net: split channel-based driver impl from usb cdc-ncm into a separate crate.
